# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  schemas.CreateLearnedWordRequest:
    properties:
      cnt_reviewed:
        type: integer
      confidence_score:
        type: integer
      last_reviewed:
        type: string
      learned_at:
        type: string
      user_id:
        type: string
      word_id:
        type: string
    required:
    - user_id
    - word_id
    type: object
...
  schemas.ErrorResponse:
    properties:
      message:
        example: invalid request
        type: string
    type: object
  schemas.LearenedWordResponse:
    properties:
      cnt_reviewed:
        type: integer
      confidence_score:
        type: integer
      last_reviewed:
        type: string
      learned_at:
        type: string
      user_id:
        type: string
      word_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: Woolfer0097@yandex.ru
    name: Danila Kochegarov
    url: http://fluently.com
  description: Backend API for Fluently Telegram bot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://fluently.com/terms/
  title: Fluently API
  version: "1.0"
paths:
...
  /users/{user_id}/learned-words/:
    get:
      consumes:
      - application/json
      description: Returns all learned words for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.LearenedWordResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Get user's learned words
      tags:
      - learned-words
    post:
      consumes:
      - application/json
      description: Adds a new learned word for a user
      parameters:
      - description: Learned word data
        in: body
        name: learned_word
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateLearnedWordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Add a learned word
      tags:
      - learned-words
  /users/{user_id}/learned-words/{word_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's learned word by word ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Word ID
        in: path
        name: word_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Delete a learned word
      tags:
      - learned-words
    get:
      consumes:
      - application/json
      description: Returns a learned word by user ID and word ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Word ID
        in: path
        name: word_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.LearenedWordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Get user's learned word by word ID
      tags:
      - learned-words
    put:
      consumes:
      - application/json
      description: Updates a user's learned word by word ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Word ID
        in: path
        name: word_id
        required: true
        type: string
      - description: Learned word data
        in: body
        name: learned_word
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateLearnedWordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Update a learned word
      tags:
      - learned-words
...
swagger: "2.0"